add function about privacy policy and key back
